+++ services/limbless-app/limbless-db/limbless_db/models/SampleAttribute.py
@@ -11,7 +11,7 @@ class SampleAttribute(Base):
     sample_id: Mapped[int] = mapped_column(sa.Integer, sa.ForeignKey("sample.id"), nullable=False, primary_key=True)
     name: Mapped[str] = mapped_column(sa.String(64), nullable=False, primary_key=True, index=True)
     value: Mapped[str] = mapped_column(sa.String(64), nullable=False)
-    type_id: Mapped[int] = mapped_column(sa.Integer, nullable=False)
+    type_id: Mapped[int] = mapped_column(sa.SmallInteger, nullable=False)
 
     @property
     def type(self) -> AttributeTypeEnum:
